requirements.txt
Pillow==10.3.0
rembg==2.0.50 